extracted type-support in own package
