+++ .gitignore
@@ -11,3 +11,4 @@ build/
 .gradle/
 .idea/
 ca.iml
+/bin/
