+++ src/main/java/org/epics/ca/impl/Messages.java
@@ -4,7 +4,7 @@ import java.net.InetAddress;
 import java.nio.ByteBuffer;
 
 import org.epics.ca.Constants;
-import org.epics.ca.impl.TypeSupports.TypeSupport;
+import org.epics.ca.impl.types.TypeSupports.TypeSupport;
 import org.epics.ca.util.net.InetAddressUtil;
 
 public final class Messages {
